correction bug et cacher valider si aucune demande
