+++ App/ProtoControllers/Responsable/Traitement/Conge.php
@@ -80,7 +80,9 @@ class Conge extends \App\ProtoControllers\Responsable\ATraitement
         ob_start();
         $table->render();
         $return .= ob_get_clean();
-        $return .= '<div class="form-group"><input type="submit" class="btn btn-success" value="' . _('form_submit') . '" /></div>';
+        if (!empty($demandesResp) && !empty($demandesGrandResp) ) {
+            $return .= '<div class="form-group"><input type="submit" class="btn btn-success" value="' . _('form_submit') . '" /></div>';
+        }
         $return .='</form>';
 
         return $return;
@@ -138,9 +140,9 @@ class Conge extends \App\ProtoControllers\Responsable\ATraitement
         $infoDemandes = $this->getInfoDemandes(array_keys($put['demande']));
 
         foreach ($put['demande'] as $id_conge => $statut) {
-            if (\App\ProtoControllers\responsable::isRespDeUtilisateur($resp, $infoDemandes[$id_conge]['p_login'])) {
+            if (\App\ProtoControllers\Responsable::isRespDeUtilisateur($resp, $infoDemandes[$id_conge]['p_login'])) {
                 $return = $this->putResponsable($infoDemandes[$id_conge], $statut, $put, $errorLst);
-            } elseif (\App\ProtoControllers\responsable::isGrandRespDeGroupe($resp, \App\ProtoControllers\Utilisateur::getGroupesId($infoDemandes[$id_conge]['p_login']))) {
+            } elseif (\App\ProtoControllers\Responsable::isGrandRespDeGroupe($resp, \App\ProtoControllers\Utilisateur::getGroupesId($infoDemandes[$id_conge]['p_login']))) {
                 $return = $this->putGrandResponsable($infoDemandes[$id_conge], $statut, $put, $errorLst);
             } else {
                 $errorLst[] = _('erreur_pas_responsable_de') . ' ' . $infoDemandes['id_conge']['p_login'];
@@ -163,7 +165,7 @@ class Conge extends \App\ProtoControllers\Responsable\ATraitement
                 $return = $this->updateStatutRefus($id_conge, $put['comment_refus'][$id_conge]);
                 log_action($infoDemande['p_num'], 'refus', '', $infoDemande['p_login'], 'traitement demande ' . $id_conge . ' (' . $infoDemande['p_login'] . ') (' . $infoDemande['p_nb_jours'] . ' jours) : refus');
             } elseif (\App\Models\Conge::ACCEPTE === $statut) {
-                if (\App\ProtoControllers\responsable::isDoubleValGroupe($infoDemande['p_login'])) {
+                if (\App\ProtoControllers\Responsable::isDoubleValGroupe($infoDemande['p_login'])) {
                     $return = $this->updateStatutPremiereValidation($id_conge);
                     log_action($infoDemande['p_num'], 'valid', $infoDemande['p_login'], 'traitement demande conges ' . $id_conge . ' de ' . $infoDemande['p_login'] . ' première validation');
                 } else {
@@ -190,7 +192,7 @@ class Conge extends \App\ProtoControllers\Responsable\ATraitement
                 $return = $this->updateStatutRefus($id_conge, $put['comment_refus'][$id_conge]);
                 log_action($infoDemande['p_num'], 'refus', '', $infoDemande['p_login'], 'traitement demande ' . $id_conge . ' (' . $infoDemande['p_login'] . ') (' . $infoDemande['p_nb_jours'] . ' jours) : refus');
             } elseif (\App\Models\Conge::ACCEPTE === $statut) {
-                if (\App\ProtoControllers\responsable::isDoubleValGroupe($infoDemande['p_login'])) {
+                if (\App\ProtoControllers\Responsable::isDoubleValGroupe($infoDemande['p_login'])) {
                     $return = $this->putValidationFinale($id_conge);
                     log_action($infoDemande['p_num'], 'ok', $infoDemande['p_login'], 'traitement demande ' . $id_conge . ' (' . $infoDemande['p_login'] . ') (' . $infoDemande['p_nb_jours'] . ' jours) : OK');
                 } else {
@@ -360,13 +362,13 @@ class Conge extends \App\ProtoControllers\Responsable\ATraitement
       */
     protected function getIdDemandesResponsable($resp)
     { 
-        $groupId = \App\ProtoControllers\responsable::getIdGroupeResp($resp);
+        $groupId = \App\ProtoControllers\Responsable::getIdGroupeResp($resp);
         if (empty($groupId)) {
             return [];
         }
 
         $usersResp = [];
-        $usersResp = \App\ProtoControllers\responsable::getUsersGroupe($groupId);
+        $usersResp = \App\ProtoControllers\Responsable::getUsersGroupe($groupId);
         if (empty($usersResp)) {
             return [];
         }
@@ -389,12 +391,12 @@ class Conge extends \App\ProtoControllers\Responsable\ATraitement
       */
     protected function getIdDemandesGrandResponsable($gResp)
     {
-        $groupId = \App\ProtoControllers\responsable::getIdGroupeGrandResponsable($gResp);
+        $groupId = \App\ProtoControllers\Responsable::getIdGroupeGrandResponsable($gResp);
         if (empty($groupId)) {
             return [];
         }
         
-        $usersResp = \App\ProtoControllers\responsable::getUsersGroupe($groupId);
+        $usersResp = \App\ProtoControllers\Responsable::getUsersGroupe($groupId);
         if (empty($usersResp)) {
             return [];
         }
